requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
requests==2.32.3
python-dateutil==2.9.0.post0
pytest==8.3.5
playwright==1.51.0
nodriver==0.39
pyvirtualdisplay==3.0
